Setup up value access for parameters of UIJson class.
