+++ geoh5py/ui_json/ui_json.py
@@ -19,6 +19,7 @@ from __future__ import annotations
 
 from typing import Any
 
+from geoh5py.ui_json.descriptors import ValueAccess
 from geoh5py.ui_json.enforcers import EnforcerPool
 from geoh5py.ui_json.forms import FormParameter
 from geoh5py.ui_json.parameters import Parameter
@@ -45,6 +46,14 @@ class UIJson:
         self.enforcers: EnforcerPool = EnforcerPool.from_validations(
             self.name, self.validations
         )
+        self._allow_values_access()
+
+    def _allow_values_access(self):
+        """Public parameter names access underlying parameter value."""
+        for parameter in self.parameters:
+            if parameter not in dir(self):
+                setattr(self, f"_{parameter}", self.parameters[parameter])
+                setattr(self.__class__, parameter, ValueAccess(f"_{parameter}"))
 
     def to_dict(self, naming: str = "snake") -> dict[str, Any]:
         """
@@ -70,7 +79,6 @@ class UIJson:
 
     def validate(self):
         """Validates uijson data against a pool of enforcers."""
-
         uijson = self.to_dict()
         self.enforcers.enforce(uijson)
 
@@ -83,6 +91,6 @@ class UIJson:
         if "title" in uijson:
             name = uijson["title"]
         elif "geoh5" in uijson:
-            name = uijson["geoh5"].h5file
+            name = uijson["geoh5"].h5file.stem
 
         return name
